update: simplifier codebase
