+++ madara/crates/client/gateway/client/src/tests/mod.rs
@@ -1,1 +1,1 @@
-pub mod retry_stress_tests;
+pub mod retry_tests;
